Merged PR 15034: 19907217 [Meriwether] Stub generation selects wrong version of API, resulting in content drop-out
There were multiple versions of stubs created for some apis. The ones that've I've added/replaced have the correct content. The alternative versions end in ~r1, 2, 3 … In some cases, these replace the current topic (and include all the required content.) In others, the topic is in addition to the current (correct) content.