Merged PR 15324: Fix links in FindResourceA and FindResourceExA pages
These have links to a page to describe all the resource types, but the links go to the D3D resource reference, not the Win32 idea of a resource. This change fixes the links to go to an appropriate Win32 reference.
This is for https://github.com/MicrosoftDocs/feedback/issues/1203