Merged PR 15491: Fix text and hyperlink in page for D2D1_RENDERING_CONTROLS
This is for addressing https://github.com/MicrosoftDocs/feedback/issues/1640 .

The text was written in a convoluted way with the text "Output.BufferPrecision" linking to CreateGradientStopCollection's buffer precision. In this context I think the text was supposed to be about setting the buffer precision property of effect nodes in the graph. This change removes the misleading link and fixes the text.