Merged PR 15935: Improve parameter documentation for ClearUAV
Improve parameter documentation for ClearUAV. @<Steven White> feel free to re-word as you feel necessary. The key is to remove the misinformation ("the start of the heap" and references to render targets instead of UAVs) and add the fact that the two descriptors have to come from different heaps, where the shader-visible one must be bound to the command list.