Merged PR 15937: Add cache lookup details for isolated name translation.
Add cache lookup details for isolated name translation. This change only fixes the LsaLookupNames and LsaLookupNames2 documentation.  The bulk of the new content is centralized in LsaLookupNames2, with LsaLookupNames now having a link to LsaLookupNames2.

Once we are agreed on this approach, we will also add links to LsaLookupNames2 from LookupAccountName[A|W].