Document the EOF behaviour for synchronous handles with an OVERLAPPED struct
The documentation doesn't seem to cover the EOF behaviour when a synchronous file handle is accessed with an OVERLAPPED struct.  That is a useful thing to do, in order to provide an offset, to simulate the POSIX function pread() (ie to avoid having a separate lseek() syscall).  See also Raymond Chen's blog[1].

[1] https://devblogs.microsoft.com/oldnewthing/20150121-00/?p=44863