fileapi: Escape backslashes for \\?\
Prior to this change the MS documentation for certain API functions
showed incorrect "\?" for the extended-length path prefix instead of
the correct "\\?\", because the backslashes were not escaped.

This is only a partial fix, this is as much as I can do. There are
probably many more functions that need to have \\?\ escaped in certain
areas. Also \\.\ probably needs to be escaped as well.

Closes #xxxx