Merged PR 16162: FormatMessage: Don't document internal behavior of LocalFree
FormatMessage: Don't document internal behavior of LocalFree

Was introduced when we were considering how to enable `FORMAT_MESSAGE_ALLOCATE_BUFFER` in the Store, and one option was documenting enough of the internals of LocalFree so you could free the buffer.

We decided not to go with that approach and instead add LocalFree to WACK.

This documentation was apparently checked in before the final decision was made not to document the internals of LocalFree. (It's also apparently a copy-paste from internal developer notes and wasn't cleaned up for publication. The wording is confusing and it talks about counterfactuals.)