Merged PR 16235: Add a note in the doc for LCMapStringEx() that Sort keys are opaque byte streams.
Add a note in the doc for `LCMapStringEx()` that Sort keys are opaque byte streams.

I didn't add the same note in the doc for `LCMapStringA` and `LCMapStringW` because in their Remarks sections, they both point to the Remark section of `LCMapStringEx()`.

Earlier, I tried making this change in the GitHub sdk-api repo. Found out after that I should use cpubwin instead. So, I've closed this PR: https://github.com/MicrosoftDocs/sdk-api/pull/111