Specify LoadLibraryExW dotted filename behavior
Specify the behavior of LoadLibraryExW() when passed a DLL module name without a path and `.dll` extension, to highlight the fact that the `.dll` extension is only appended automatically to the module name if that module base name does not already contain some dot `.` character. This is a source of confusion which leads to a bug in `DllImport` (https://github.com/dotnet/runtime/issues/7223).