Merged PR 16441: fixing ref parms, etc.
This was mostly missing parameter descriptions, redundant parameters, and incorrect types (for example, pointer missing or reference missing or one in place of the other). SAL not being rendered, in rare cases not being rendered and also being incorrect. A lot more of these fixes to come over the next day or so.