Merged PR 16522: PF_VIRT_FIRMWARE_ENABLED requires availability as well
PF_VIRT_FIRMWARE_ENABLED requires availability as well

The feature could be enabled in firmware, but not available to the OS. For example, the OS could be running inside a virtual machine, so the virtualization extensions are not available to the OS because the host OS is using them. Note that if Hyper-V is enabled, then the root OS is itself running inside a virtual machine (the hypervisor is the host OS).

See https://stackoverflow.microsoft.com/questions/172196 for details.