WinUser.h: user32.dll ! SetForegroundWindow minor technical corrections
The SetForegroundWindow doc page remarks were not quite correct
about the cases when a process can call the function to
set the foreground window.

This change makes some corrections to the doc page.
However, the SetForegroundWindow rules are very complex,
so a full accounting of when you can set foreground
would probably be too long to fit - and would encode
details that might be misused to annoy the user.
I thus added a generic escape clause to note that
there may be additional restrictions not stated.