Clarify SetProcessValidCallTargets documentation
Getting this function to work in my scenario was a bit of a headbanger. The root cause of my problems where that I use W^X: I allocated the page as read and write only, generate the code and then switch to read and execute. This means Windows didn't have a CFG bitmap for the memory page, causing this method to fail.

Hopefully these changes should avoid other people falling in that trap.

Thanks to @riverar for the assistance in diagnosing this issue.