Merged PR 16847: Update sdk-api master branch with f1_keywords from stubs/master branch
It is intended to update all f1_keywords values. Meanwhile, it also updates the format of markdown files. I updated such files with the newly developed mwyamldiff tool. Thus, the changes are not subject to manual typos. Besides, the header files (index.md) are also updated by the tool although they don't need f1_keywords, which looks OK to update. In a word, the changes are safe.

Related Work Item:
https://dev.azure.com/ceapex/Engineering/_workitems/edit/291909