Clarify this function requires filesystem support
This function relies on filesystem support. If a filesystem does not support it, ERROR_INVALID_PARAMETER is returned.

Also, the type of the streams this function affects can best be described here as ':$DATA' (technically NTFS defines ':' as the delimiter and '$DATA' as the stream type, for the final construction of `filename:streamname:streamtype`; for those who don't work much with filesystems, though, it seems a rather academic distinction).  '::$DATA' refers to the unnamed (default) data stream on a file.  Directories cannot have an unnamed data stream, so referring to "the first ::$DATA stream type" is incorrect.

Thanks for working as hard as you do on making sure we can understand what your company has created for us to use!