Convert HTML to Markdown (PM_COLLECT_PROC)
Fixed a lot of formatting wrongdoings, as well as renamed
and reordered parameters consistently with callback signature.

Two parameters, namely *lpValueName* and *lpNumObjectTypes* were
annotated incorrectly in source code (or at least the way they were
does not make sense), so a [!NOTE] were added to each one to mitigate
that oversight.