Change links to point to *W version
I don't know what the policy for cross-links to other APIs and data structures is: do articles that describe the unicode behavior (W suffix) always link to the ASCII versions of other APIs (A suffix)? Or do unicode APIs link to other unicode APIs? If the latter, then this change is changing all the links in this article to the unicode versions.