Proposed additions to WTS_CONNECTSTATE_CLASS
While working with `WTS_CONNECTSTATE_CLASS` recently, both I and a colleague had difficulty interpreting all of the possible enumeration values based on this documentation. Through experimentation, we ascertained the information I've now added via this PR.

I think it would be great to add these clarifications to the documentation, but I also have some reservations, so I'm really sending this for a discussion. In particular, I'm not 100% satisfied with the following:

- I can't tell if the information is complete. I might have uncovered certain pieces of information, while inadvertently excluding others.
- The details are only being added for the two values I examined. It would be great to include similar details for the other values.

Ideally, it would be great if we could loop a Subject Matter Expert into this review as well.

I'm happy to make whatever changes we decide on, or to shelve this review entirely if favoured.