Mention ERROR_INSUFFICIENT_BUFFER
Documentation did not mention that ERROR_INSUFFICIENT_BUFFER is the error code when buffer is too small. Although that would be obvious to anyone familiar with Win32 APIs, mentioning it explicitly will help avoid any doubt and help newcomers.

Also mention that you are allow to pass NULL pointers for the buffers if the buffer size is being passed as zero.