Update nf-cfapi-cfsetinsyncstate.md
The InSyncUsn parameter does not work. The operation will fail using anything but NULL (including the correct USN).

After the method returns it will not write the new USN to the paramater in any case.

This is demonstrated by [this sample](https://github.com/LokiMidgard/CloudTest).

There was also a [stackoverflow](https://stackoverflow.com/questions/60816543/cfsetinsyncstate-does-not-behave-as-documented-cfapi) discussion, but this did not lead to any results. (It desscribed a workaround to get the USN from a file but not to run the operation conditionally).