Merged PR 17218: Clarify treatment of null lpSubKey in RegOpenKeyEx*
Existing documentation said that you cannot pass NULL for predefined keys. This has not been true since Windows NT 3.5,which added the lines

```cpp
    //
    //  If lpSubKey is NULL, then assume NUL-string as subkey name
    //

    if( lpSubKey == NULL ) {
        NullString = UNICODE_NULL;
        lpSubKey = &NullString;
    }
```

When this code was added, we neglected to take the opportunity to clean up the rest of the code that deals with `lpSubKey`, so the rest of the code still applies the old policies for `lpSubKey = NULL` even though it's not possible any more.

Since we don't generally document Windows behavior prior to Windows XP, there's no need to call out the behavior change in Windows NT 3.5. Just act like it's always been this way.

Bonus 1: Document the key refresh behavior, which is significant for HKEY_CURRENT_USER.

Bonus 2: Predefined keys are excluded from transactions. This was not called out in RegOpenKeyTransacted.

```cpp
    //
    // predefined handles *cannot* be opened in transactional context
    //
    OldTxHandle = RtlGetCurrentTransaction();
    RtlSetCurrentTransaction(NULL);
```