Merged PR 17241: Clarify returned key of RegOpenKeyEx* when lpSubKey is NULL or empty
Previous text added in !17218 was poorly phrased: "If A and B, then behavior 1. Otherwise, behavior 2." This was intended to mean "Otherwise, if A but not B, then behavior 2." However, what it actually said was "In all other cases, then behavior 2."

Rephrase so that behavior 2 is correct for all other cases: The returned key is a unique key.

The only time the returned key is a copy of the inbound key is if lpSubKey is NULL/empty and hKey is paredefined key.