LoadStringA can't return string resource directly
It's actually impossible for the ANSI version of LoadString to be
implemented as documented, because the string resources are stored
in Unicode only.

If you try to call LoadStringA with cchBufferMax=0, it will attempt to
null-terminate the 0-length buffer (i.e. it will do `lpBuffer[-1] = 0`,
corrupting nearby memory), and return -1.