Merged PR 17374: Clarify treatment of lpSubKey in RegSetKeyValue
Existing text said that if the subkey is NULL, then "the key gets a default security descriptor," which is incorrect. If the subkey is NULL, then we don't change the key security.

Explained what happens if the subkey is non-NULL: It is created as a non-volatile key with default security.