Merged PR 17382: IRpcChannelBuffer::IsConnected return value documented backward
The function returns HRESULT, so it's S_OK for "connected" and S_FALSE for "disconnected". Numerically, this is *opposite* of the TRUE/FALSE values given in the documentation. (S_OK = FALSE and S_FALSE = TRUE.) So anybody who followed the documentation and wrote

```cpp
if (rpcChannelBuffer->IsConnected() == FALSE) ...
```

had their test backward.