Merged PR 17585: Fix double UTF-8 encoding
Particularly amusing is that the double--utf8 encoding in mmioOpen* went through entity escaping, so the en-dash was UTF-encoded to E2 80 93, and then the 80 got encoded as `&euro;`!

Also fix use of fancy punctuation (often by removing it). The fancy punctuation marks were sometimes used incorrectly or confusingly. For example, all of the VFW functions used en-dash instead of minus sign.

The two parts are in separate commits, so you can validate them separately.