Update ns-wingdi-displayconfig_path_info.md
Request for clarification or reference link.

Also, there used to be a *Submit Feedback (about) This Page* button at the bottom of these pages. I'm glad that I can now make direct edits and submit an MR when I have a specific fix to propose. However, it's not clear to me what the process is now for requesting *additional information*, and adding my comments to the doc and submitting an MR feels like misuse -- please let me know if that's the case, and what the intended process is. Thanks.