Update ne-wingdi-displayconfig_device_info_type.md
Requests to fill out documentation for three enumerands and add pages for two struct types (see comments inline).

Also, there used to be a button at the bottom of many docs.microsoft.com pages to generate a GitHub issue with links back to the docs page. That seems to be gone now, and it's not clear what your preferred procedure is for us to request doc changes that we can't entirely edit & submit on our own. Please let me know, and I'll use that going forward vs. creating MRs.