Merged PR 18139: Message fixes | Validation issues | 5 (BULK)
This is a bulk update.

This PR is to fix issues with disallowed HTML attributes.

If these changes are not appropriate, please let me know. I'm only trying to clean up validation issues in this repo. So I can easily close this PR or modify it in whatever way is best for the content. I'll close this PR after one business day if there is no feedback.