Update nf-netfw-networkisolationenumappcontainers.md
We had a software vendor ask about this, and it isn't clear to me from the current doc description what to do.  Looking at our current implementation, it appears that NetworkIsolationFreeAppContainers would be the correct call for freeing the memory.  Not freeing it in any way results in a memory leak.