Clarify that Security Advisory 2915720 is opt-in
WinVerifyTrust and WinVerifyTrustEx functions (wintrust.h) have a
note regarding WinVerifyTrust Signature Validation Vulnerability
(CVE-2013-3900). This note leads to Security Advisory 2915720 which
prompts administrators to configure `EnableCertPaddingCheck` registry
key for additional validation.

The WinVerifyTrust{,Ex} function documentation states that registry
key value will be set to "1" by default "on June 10, 2014" however
this is not correct. The Advisory linked was amended, as to initially
push back the defaults change, and then amended (V1.4) on July 29, 2014
as to clarify there are no active plans to set "1" as default.

These documents are being updated, as to avoid conflicting information
and potential resulting confusion among system administrators, who are
reading Microsoft documentation to understand the change.