filesys flag table: fix malformed definition lists
The Value column of the filesystem flag table was using HTML definition
lists (<dl>) for the identifier and value. However, the definition lists
have only a single item where both fields were tagged as <dt>
(definition term), and there was no <dd> (definition) body.

Changed these to just bold identifier, linebreak, and normal text for
the flag value. This is in line with the content that appears post
build.