Update the connectorInstance field description
'type' in this context is too vague. It seems that it may be referring to `DISPLAYCONFIG_TARGET_BASE_TYPE.baseOutputTechnology`?

We were trying to use `connectorInstance` as part of a composite key that would uniquely identify each target monitor connected to an adapter. However, a test system with three monitors connected returned **0** for each of them. The same system with two of those monitors disconnected (leaving only the primary monitor) returned **1** for connectorInstance.

**`connectorInstance`** seems to need a clearer, more comprehensive explanation.

I think it would also be very helpful to identify the fields (from various CCD structs) that, taken together, uniquely identify a target device, and are stable between OS restarts. It seems to me to be a common requirement to identify a display, associate some data/state to it, persist that data between runs of an app, and then retrieve it and use it to locate that display again (if it's connected). We thought connectorInstance would be a good way to do that, but it seems not. Since Microsoft's **Display settings** app seems to do all of this, some explanation would be very helpful and avoid a lot of empirical experimentation...