Merged PR 18503: Be less specific about how threadpool works in WaitForMultipleObjects(Ex)
Existing documentation for WaitForMultipleObjects(Ex) provides suggestions for waiting on more than 64 objects. That's good. However, the suggestion to use the thread pool goes into implementation details which are no longer true. The thread pool now uses an I/O completion port to multiplex the handle waits, rather then grouping the waits into blocks of 63.

Changed the description to be less specific. We don't say that the thread pool groups waits into blocks of 63; we just say that it waits efficiently, and the implementation figures out the most efficient way to do it.