Merged PR 18514: Attempt to fix the missing includes in MCCS documentation
The `!include` tags still weren't working on our MCCS documentation. I looked carefully at other examples of it working in the Windows docs repos, and I think it's because I wrote it as `[!INCLUDE [title](url)]` instead of `[!include[title](url)]`, lowercase and with no space. The VS Code extension accepts any of these, but perhaps the reference docs repo only accepts the lowercase version?

Update: Lori pointed out that the issue could be that the `!include` is between the H1 and the "description", so I've put them into the description. This feels a lot more like the real issue, but unfortunately I don't know how to test it beyond just checking it in and seeing what we get on the site.