Added C++ example code.
I faced many issues, so many would have faced, that, the LPVOID is not actually required for storing the gamma values. The lpRamp variable can be a WORD Array of 3x256 elements. We can just pass the array into the function, and it will update it.
This example will save a lot of people's time and hardwork.