Clarify LeaveCriticalSection guarantees
In sdk-api-src/content/synchapi/nf-synchapi-deletecriticalsection.md the following was stated:
> The caller is responsible for ensuring that the critical section object is unowned and the specified CRITICAL_SECTION structure is not being accessed by any critical section functions called by other threads in the process.

This puts a note on a CRITICAL_SECTION structure access, but it was never stated, how this structure is accessed by LeaveCriticalSection.

The proposed change ensures that this structure is only accessed until a critical section object ownership is released.