Update nf-synchapi-sleepex.md
In the description of the dwMilliseconds parameter, the mention of pending APCs combined with the bAlertable parameter set to FALSE was misleading.  When bAlertable is FALSE, APCs will not be delivered.  I.e. the result is the same whether an APC is queued or not - no APC callback will execute.  To say "if there are no pending user APCs" implied that this condition would somehow impact the outcome.

In the bAlertable description, I was initially confused by the phrase "the function does not return".  I now understand that it means "does not return immediately".  I added those words for clarity, but leave it at your discretion as to whether that is needed.

Small typo correcting "timer-out" to "time-out"