Merged PR 19029: GetDistanceOfClosestLanguageInList corrections
The special case of wchListDelimiter = L'0\' is not documented.

Instructions on interpreting the pClosestDistance were missing.

And in fact even though the function is called "GetDistance", the returned "distance" is not a distance at all! It is a "score".

You want to minimize distance, and an exact match would have a distance of 0.

But this function returns 0.0 to mean "no match" and 1.0 to mean "exact match". You want to maximize the resulting value, not minimize it.