Fixed wrong link under the style field (#1327)
The description of the style field implies that the link will take you to the page which has a list of the Window Class Styles. However, currently, it does not. Instead, it sends you to a general page about window classes (https://learn.microsoft.com/en-us/windows/win32/winmsg/about-window-classes). This can be very confusing for someone who is new to Win32 and was expecting a list of the styles. This pull request should fix that issue and thus make it easier to use this page.