Update the document to reflect actual annotations (#1168)
In the library header winver.h, this function has _Reserved_ annotation on the third "dwHandle" parameter, requiring the value to be NULL or 0(zero). But currently this document indicates that the value will be ignored instead. Either update the document to reflect the current annotation on the parameter, or have the owner of the header to update the annotation to reflect the actual behavior if the function really ignores this value.