Change parameter value to avoid data type warning
The parameter `hProv` is of data type `HCRYPTPROV_LEGACY` which is `ULONG_PTR` which is `unsigned long long`. Thus when passing `NULL` as `hProv` to `CertOpenSystemStoreA` or `CertOpenSystemStoreW`, this will create a warning regarding incompatible pointer to integer conversion. Since `NULL` is just zero casted to `void *`, passing zero is an acceptable substitution.