Merged PR 20454: The ISoftwareBitmapNativeFactory methods create a SoftwareBitmap
Documentation said that they created an ISoftwareBitmapNative, which is an interface, not an object. Also, you rarely want the ISoftwareBitmapNative anyway. You usually want the ISoftwareBitmap.

Also fix a copy/pasta in the Note. The IWICBitmap method still talked about the MF buffer.

The other half of this fix is !20455