Merged PR 20559: Clarify behavior if waitable timer elapses twice while APC still pending
What happens is that the second APC is not queued if there is already an APC pending. APCs do not accumulate; they coalesce.

Existing text was confusing. It said "If the timer is set before the thread enters an alertable wait state, the APC is canceled." Which suggests that you have to hurry up and do a `SleepEx(0, TRUE)` after setting the timer to "enable" the APC before the timer elapses. That is not what the text is trying to say. It's saying that if the thread fails to enter an alertable wait state for so long that the timer wants to queue two APCs, then the second APC is canceled. Except that is *still* wrong, because the APC is not canceled; it never existed!

I hope the new text is clearer.