Fix typo in nf-winnls-lcmapstringex.md
In the "Return value" the first paragraph already talks about what is returned when the function is used for string mapping. The second paragraph intended to say what is returned when the function is used to generate a sort key, but used the words for the string mapping case again.