Update nn-wtsprotocol-iwrdsprotocolconnection.md
Updated the contract to make sure it is clear that if the work is moved to a new thread that the creator of the new thread cannot wait on the thread to complete the work. The entire purpose of making a new thread to do the work is to make the work async, non-blocking.