Merged PR 20861: Clarify overwrite means truncate
The `CREATE_ALWAYS` flag was documented as "If the file exists, the function overwrites the file." But this is confusing, because *all* the open modes will overwrite the file when you write to it. The thing that makes `CREATE_ALWAYS` special is that it *truncates* the file: The file size is reset to zero.