Patch 1 FindWindowEx A/W (#1630)
* Parameter Type LPCSTR

In the List of Parameters the Type of lpszClass is listed as LPCTSTR, although this is only the case for FindWindowEx, not for FindWindowExA or FindWindowExW. 
Btw. The pages for FindWindowExA and FindWindowExW are exactly identical except for the gray Syntax Box

* Parameter type updated to LPCWSTR

In the List of Parameters the Type of lpszClass is listed as LPCTSTR, although this is only the case for FindWindowEx, not for FindWindowExA or FindWindowExW. 
Btw. The pages for FindWindowExA and FindWindowExW are exactly identical except for the gray Syntax Box