Merged PR 20823: Mark SHCONTF_FLATLIST as obsolete
It was introduced in shell:44967 for the 2003 Longhorn PDC and was implemented in dbfolder.cpp (CDBFolder::EnumObjectsEx).

It never shipped in Vista.

The code finally RI'd into Win7 but the support was deleted in shell:501772 in 2009 with the comment "issues in nearly dead code", so it never shipped in any RTM vehicle.

It was "dead on arrival".

The current documentation [is a source of confusion](https://stackoverflow.com/q/77303032), so I'm just changing it to say "Obsolete. Do not use."

Also fix typo that said "SHCONTF was renamed to SHCONTF".