winuser: Clarify wsprintfA and wsprintfW format specifiers
- Explain how format specifiers are treated differently depending on
  whether wsprintfA or wsprintfW is used.

- Explain in the format specifier section that wsprintf is a macro that
  is defined as wsprintfA (Unicode not defined) or wsprintfW (Unicode
  defined).

This is also a sync of the format specifier section in wsprintfA and
wsprintfW.

---

Before, from wsprintfW:

`s` String. This value is interpreted as type LPWSTR when the calling
application defines Unicode and as type LPSTR otherwise.

After, from wsprintfW:

`s` String. This value is interpreted as type LPSTR by wsprintfA and
type LPWSTR by wsprintfW. Note wsprintf is a macro defined as wsprintfA
(Unicode not defined) or wsprintfW (Unicode defined).

---

Prior to this change it was unclear and IMO incorrect to say that %s is
treated as LPSTR when the application does not define Unicode, because
that is documented on the wsprintfW page and wsprintfW always treats %s
as LPWSTR.

Ref: https://learn.microsoft.com/en-us/windows/win32/api/winuser/nf-winuser-wsprintfa
Ref: https://learn.microsoft.com/en-us/windows/win32/api/winuser/nf-winuser-wsprintfw

Closes #xxxxx