Merged PR 20956: ScatterGather FILE_SEGMENT_ELEMENT count incorrectly documented
ReadFileScatter and WriteFileGather do not look for a null terminator to determine how many elements to use. They infer the number of elements from the length.

Source code says:

```cpp
// NtReadFileScatter

    elementCount = BYTES_TO_PAGES(Length);
    ...
    ProbeForRead( SegmentArray,
                  elementCount * sizeof(FILE_SEGMENT_ELEMENT),
                  TYPE_ALIGNMENT(FILE_SEGMENT_ELEMENT)
                  );

    if (Length != 0) {
        //
        // Capture the segment array so it cannot be changed after
        // it has been looked at.
        //

        capturedArray = ExAllocatePool2( POOL_FLAG_PAGED | POOL_FLAG_USE_QUOTA | POOL_FLAG_RAISE_ON_FAILURE,
                                         elementCount * sizeof(FILE_SEGMENT_ELEMENT),
                                         IO_SEGMENT_ARRAY_POOLTAG );

        RtlCopyMemory( capturedArray,
                       SegmentArray,
                       elementCount * sizeof(FILE_SEGMENT_ELEMENT)
                       );

        SegmentArray = capturedArray;

// NtWriteFileGather contains identical code
```

Since the code sequence ends with an assignment to `SegmentArray`, the original `FILE_SEGMENT_ELEMENT` buffer is no longer used past this point. There is no check for a null terminating element.

This is confirmed by experimentation: I put a `FILE_SEGMENT_ELEMENT` at the end of a page and asked for exactly one page of data to be read/written. The I/O succeeded. There was no attempt to verify the presence of a null terminator.

Also clean up the documentation for `FILE_SEGMENT_ELEMENT`:
* Typo "segement"
* Fix wording "the page alignment of the data", which makes no sense.