winsock: Fix WSAStartup parameter description
- Replace wVersionRequired with wVersionRequested.

Prior to this change wVersionRequired had a description of TBD and
wVersionRequested had the correct description though was not rendered.

Both are different names of the same parameter depending on which
SDK is used. The function prototype is extracted from an SDK which uses
wVersionRequired but IMO "required" is the wrong name to use. It is
properly noted as the "requested" version numerous times throughout the
documentation.

This change makes the tradeoff that rather than be consistent with what
name the renderer has extracted from the SDK (ie wVersionRequired),
instead use the name that is most correct (ie wVersionRequested).

Recent SDKs (including the latest) and even an old Windows 7 SDK have
wVersionRequested. I couldn't find which SDK has wVersionRequired.

A complete fix for this issue would be to get the renderer to extract
the function prototype from the latest version of the SDK.

Ref: https://learn.microsoft.com/en-us/windows/win32/api/winsock/nf-winsock-wsastartup

Closes #xxxxx