Clarify behavior of relative due time in waitable timers
The relative due time skips the time spent in low-power states, at least
on Windows 10 22H2 and on Windows 11. This was somewhat surprising to me
and my colleagues, hence the suggestion to document it explicitly.

I can't confirm if it's really the case that the behavior of relative
timers changed between Windows 7 and Windows 8. However, a significant
number of Win32 APIs with relative timeouts similarly changed behavior
since Windows 8 and are documented as such. Examples include
WaitForSingleObject or GetQueuedCompletionStatus. I assume that waitable
timers were affected by the same change in Windows 8 and only the docs
update was missed. I mostly copied the existing wording from the APIs
that already document this behavior.