Update nf-threadpoolapiset-setthreadpoolwait.md
An obscure error conditions that took a bit of time to understand with access to Windows sources. Folks that do not have access to sources will be confused why passing a valid handle is raising an exception. 
Constraint is hardcoded in NtAssociateWaitCompletionPacket %SDXROOT%\xbox\lnm\ntos\io\iomgr\complete.c

    DispatcherObject = ObGetAssociatedWaitObject(TargetObject);
    if ((DispatcherObject == NULL) ||
        (KOBJECT_TYPE((PKEVENT)DispatcherObject) == MutantObject)) {

        Status = STATUS_INVALID_PARAMETER_3;
        goto Exit;
    }