MapViewOfFile - Clearer doc on matching VirtualAlloc allocation granularity
The previous doc was too vague. In Windows you can allocate memory in smaller chunks than VirtualAlloc granularity (in general). However, to use this MapViewOfFile API it specifically needs to be a multiple of VirtualAlloc granularity. The same info can be found by following the links, but this makes it clearer